+++ zeta/flags/internal/registry.h
@@ -10,6 +10,7 @@
 
 #include <atomic>
 #include <cstddef>
+#include <string>
 #include <string_view>
 
 namespace zeta {
@@ -34,8 +35,10 @@ inline std::atomic<FlagEntry*> g_registry_head{nullptr};
 class FlagEntry {
 public:
     constexpr FlagEntry(const char* name, const char* help,
-                        const char* filename) noexcept
-        : name_(name), help_(help), filename_(filename), next_(nullptr) {}
+                        const char* filename, const char* env_name = nullptr,
+                        bool required = false) noexcept
+        : name_(name), help_(help), filename_(filename), env_name_(env_name),
+          required_(required), next_(nullptr) {}
 
     /// Self-register into the global linked list.  Safe to call from
     /// static constructors because `g_registry_head` is constant-init.
@@ -52,11 +55,20 @@ public:
     [[nodiscard]] std::string_view Name()     const noexcept { return name_; }
     [[nodiscard]] std::string_view Help()     const noexcept { return help_; }
     [[nodiscard]] std::string_view Filename() const noexcept { return filename_; }
+    [[nodiscard]] std::string_view EnvName() const noexcept {
+        return env_name_ != nullptr ? std::string_view(env_name_) : std::string_view{};
+    }
+    [[nodiscard]] bool Required() const noexcept { return required_; }
+    [[nodiscard]] bool IsSet() const noexcept {
+        return is_set_.load(std::memory_order_relaxed);
+    }
     [[nodiscard]] FlagEntry*       Next()     const noexcept { return next_; }
 
     virtual bool Parse(std::string_view /*value*/) noexcept { return false; }
     [[nodiscard]] virtual std::string_view CurrentValue() const noexcept { return ""; }
+    [[nodiscard]] virtual std::string_view DefaultValue() const noexcept { return ""; }
     [[nodiscard]] virtual std::string_view TypeName() const noexcept { return ""; }
+    virtual bool Validate(std::string* /*error*/) const { return true; }
 
     /// Return the head of the global registry list.
     [[nodiscard]] static FlagEntry* Head() noexcept {
@@ -64,12 +76,16 @@ public:
     }
 
 protected:
+    void MarkSet() noexcept { is_set_.store(true, std::memory_order_relaxed); }
     ~FlagEntry() = default;
 
 private:
     const char* name_;
     const char* help_;
     const char* filename_;
+    const char* env_name_;
+    bool        required_;
+    std::atomic<bool> is_set_{false};
     FlagEntry*  next_;
 };